Convert scalers to single object specific to i09
